move site context app initializer to SiteContextModule
