+++ projects/core/src/site-context/facade/language.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
-import { Store, select } from '@ngrx/store';
+import { select, Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
-import * as fromStore from '../store/index';
 import { filter, take, tap } from 'rxjs/operators';
-import { WindowRef } from '../../window/window-ref';
-import { SiteContext } from './site-context.interface';
 import { Language } from '../../model/misc.model';
+import { WindowRef } from '../../window/window-ref';
+import * as fromStore from '../store/index';
+import { SiteContext } from './site-context.interface';
 /**
  * Facade that provides easy access to language state, actions and selectors.
  */
