+++ projects/core/src/site-context/site-context.module.ts
@@ -1,12 +1,12 @@
-import { NgModule } from '@angular/core';
-import { SiteContextStoreModule } from './store/site-context-store.module';
-import { StateModule } from '../state/index';
+import { ModuleWithProviders, NgModule } from '@angular/core';
 import { Config, ConfigModule } from '../config/config.module';
+import { StateModule } from '../state/index';
 import { defaultSiteContextConfigFactory } from './config/default-site-context-config';
 import { SiteContextConfig } from './config/site-context-config';
 import { contextServiceMapProvider } from './providers/context-service-map';
 import { contextServiceProviders } from './providers/context-service-providers';
 import { siteContextParamsProviders } from './providers/site-context-params-providers';
+import { SiteContextStoreModule } from './store/site-context-store.module';
 
 // @dynamic
 @NgModule({
@@ -15,11 +15,17 @@ import { siteContextParamsProviders } from './providers/site-context-params-prov
     StateModule,
     SiteContextStoreModule,
   ],
-  providers: [
-    contextServiceMapProvider,
-    ...contextServiceProviders,
-    ...siteContextParamsProviders,
-    { provide: SiteContextConfig, useExisting: Config },
-  ],
 })
-export class SiteContextModule {}
+export class SiteContextModule {
+  static forRoot(): ModuleWithProviders {
+    return {
+      ngModule: SiteContextModule,
+      providers: [
+        contextServiceMapProvider,
+        ...contextServiceProviders,
+        ...siteContextParamsProviders,
+        { provide: SiteContextConfig, useExisting: Config },
+      ],
+    };
+  }
+}
