+++ projects/storefrontlib/src/config/index.ts
@@ -1,2 +1,1 @@
 export * from './provide-config-from-meta-tags';
-export * from './storefront-config';
