+++ projects/storefrontlib/src/storefront-config.ts
@@ -9,12 +9,12 @@ import {
   SiteContextConfig,
   StateConfig,
 } from '@spartacus/core';
-import { CheckoutConfig } from '../cms-components/checkout/config/checkout-config';
-import { IconConfig } from '../cms-components/misc/icon/index';
-import { PWAModuleConfig } from '../cms-structure/pwa/index';
-import { LayoutConfig } from '../layout/config/layout-config';
+import { CheckoutConfig } from './cms-components/checkout/config/checkout-config';
+import { IconConfig } from './cms-components/misc/icon/index';
+import { PWAModuleConfig } from './cms-structure/pwa/index';
+import { LayoutConfig } from './layout/config/layout-config';
 
-export type StorefrontModuleConfig =
+export type StorefrontConfig =
   | AuthConfig
   | CmsConfig
   | OccConfig
